refactor: move templates into registrars
